Organize ayaka-script parser impls.
